requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
django-allauth==65.2.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
idna==3.10
iniconfig==2.0.0
packaging==24.2
phonenumbers==8.13.49
pillow==11.0.0
pluggy==1.5.0
pytest==8.3.3
pytest-django==4.9.0
requests==2.32.3
sqlparse==0.5.1
stripe==11.2.0
typing_extensions==4.12.2
urllib3==2.2.3
